Renamed usercontextjs to contextjs, added a comment to login, added home button to navbar.
