+++ northark-frontend/src/components/NavBar.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom'
 import { useContext, useEffect, useState } from 'react';
-import { UserContext, SectionContext } from '../contexts/UserContext'
+import { UserContext, SectionContext } from '../contexts/Contexts'
 
 
 export const NavBar = props => {
@@ -13,12 +13,12 @@ export const NavBar = props => {
   return(
     <React.Fragment>
       <nav>
-          <h1>NorthArk</h1>
-      <div className="Navbar-Buttons">
-        <button className="Navbar-Button">Example</button>
-        <button className="Navbar-Button">Settings</button>
-        <button className="Navbar-Button" onClick={props.handleLogout}>Log Out</button>
-      </div>
+            <h1>NorthArk</h1>
+        <div className="Navbar-Buttons">
+          <button className="Navbar-Button" onClick={()=>{setSectionState("Home")}}>Home</button>
+          <button className="Navbar-Button">Settings</button>
+          <button className="Navbar-Button" onClick={props.handleLogout}>Log Out</button>
+        </div>
       </nav>
     </React.Fragment>
   )
